+++ package.json
@@ -25,8 +25,8 @@
   },
   "devDependencies": {
     "@poppanator/sveltekit-svg": "0.3.4",
-    "@sveltejs/adapter-netlify": "1.0.0-next.67",
-    "@sveltejs/kit": "1.0.0-next.367",
+    "@sveltejs/adapter-netlify": "1.0.0-next.72",
+    "@sveltejs/kit": "1.0.0-next.437",
     "@sveltejs/svelte-virtual-list": "3.0.1",
     "@types/cookie": "0.5.1",
     "@typescript-eslint/eslint-plugin": "5.30.5",
@@ -37,7 +37,7 @@
     "csstype": "3.1.0",
     "eslint": "8.19.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-svelte3": "3.4.1",
+    "eslint-plugin-svelte3": "4.0.0",
     "eslint-plugin-tailwindcss": "3.6.0",
     "husky": "7.0.4",
     "imagemin-cli": "7.0.0",
@@ -51,7 +51,7 @@
     "prettier": "2.7.1",
     "prettier-plugin-svelte": "2.7.0",
     "svelte": "3.49.0",
-    "svelte-check": "2.8.0",
+    "svelte-check": "2.8.1",
     "svelte-preprocess": "4.10.7",
     "tailwindcss": "3.1.5",
     "tslib": "2.4.0",
