--- src/global.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="@sveltejs/kit" />
-
-interface ImportMetaEnv {
-  VITE_LIFF_ID: string
-  VITE_MICROCMS_API_KEY: string
-  VITE_MICROCMS_API_ENDPOINT: string
-}
