+++ src/hooks.ts
@@ -1,10 +1,9 @@
-import * as cookie from 'cookie'
-import { v4 as uuid } from '@lukeed/uuid'
 import type { Handle } from '@sveltejs/kit'
+import * as cookie from 'cookie'
 
 export const handle: Handle = async ({ event, resolve }) => {
   const cookies = cookie.parse(event.request.headers.get('cookie') || '')
-  event.locals.userid = cookies['userid'] || uuid()
+  event.locals.userid = cookies['userid'] || crypto.randomUUID()
 
   const response = await resolve(event)
