+++ src/routes/session/[id]/+page.svelte
