+++ src/routes/speaker/[id]/+page.svelte
