+++ src/routes/todos/api.ts
@@ -1,9 +1,7 @@
 /*
 	This module is used by the /todos endpoint to
 	make calls to api.svelte.dev, which stores todos
-	for each user. The leading underscore indicates that this is
-	a private module, _not_ an endpoint — visiting /todos/_api
-	will net you a 404 response.
+	for each user.
 
 	(The data on the todo app will expire periodically; no
 	guarantees are made. Don't use it to organise your life.)
